ISAICP-6037: Move command simulation tests in its own class.
The reason is to allow 3rd party command providers to reuse this test
by supplying a different data provider.
